requirements.txt
fastapi==0.115.*
uvicorn[standard]
gunicorn
pydantic
python-dotenv
line-bot-sdk==3.17.*
tiktoken